requirements-demo.txt
commonmark==0.9.1
Pygments==2.17.2
PyYAML==6.0.1
rich==12.6.0
